add loading effect on waiting promise
