+++ src/index.html
@@ -61,7 +61,7 @@
       </form>      
     </section>
     <section class="analysis-section">
-      <ul class="link-result">                
+      <ul class="link-result">                  
       </ul>
       <div class="statistic-header">
         <h2>Advanced Statistics</h2>
