+++ src/style/link.css
@@ -52,6 +52,21 @@
   background-color: var(--DarkViolet) !important;
 }
 
+.loader {
+    border: 5px solid var(--Gray); /* Light grey */
+    border-top: 5px solid var(--DarkViolet); /* Blue */
+    border-radius: 50%;
+    width: 40px;
+    height: 40px;
+    animation: spin 1s linear infinite;
+    margin: 0 auto;
+  }
+  
+  @keyframes spin {
+    0% { transform: rotate(0deg); }
+    100% { transform: rotate(360deg); }
+  }
+
 @media screen and (min-width: 764px) {
   .link-card {
     background-color: var(--white);
